Added test logic for automatic download folder transfer to library
This is some very early logic for an automatic transfer of volumes from your download folder to your existing library. It checks what volumes are in the download folder, it finds that series in your library, it checks what volumes are in the existing library, and adds the newly downloaded volumes to the library.

I'm unsure whether all leave this logic in the script long term or not. I might cut out all logic related to user download folders in favor of it being its own script. But I haven't decided yet...