Dowload folder items are now renamed with a vXX EX: v05 instead of a v5 format
When an item is imported to the existing library, it is given a format of vXX, so v5 gets changed to v05. But the only downloaded items that are changed to this are the items that are imported, either as a completely new volume, or as an upgrade. 

When you have a completely new series that you manually copy over to your collection. Those items could be named differently, and would only get changed in the event of an upgrade. This now happens in the rename_files(). This will ensure that all downloaded items are consistent with the library before you manually move the new series over. So as long as you remember to run the script while said series is still in your download folder of course.