Alternative matching through ISBN for a downloaded volume and the existing library epub
So you have an existing library folder named "Trapped in a Dating Sim - The World of Otome Games is Tough for Mobs", and the volume you downloaded is called "Trapped in a Dating Sim". They shortened it, it won't meet the similarity requirement when automatically upgrading/moving the volume over. But, now, alternatively it will parse the downloaded file for its' ISBN, and find a matching one within your library, then proceed with the usual upgrade/move logic.