Small adjustments/renames and added support for series_id matching (see desc)
Added support for parsing a custom series_id within an epub, and in the future, a cbz file. It matches a downloaded file to a series within the user's existing library. This series_id is from google's api. This not only helps with automatic importing of new volumes that don't already exist in the library, but also volumes that don't have an associated isbn (kadakowa part releases). 

(BUT, this all relies on a script that will pull each volumes' series_id from google's api and store that metadata in the user's files. Currently that script is a WIP, and will be for quite some time. So look forward to that in the future.)