Added upgrading all volume parts to a full volume (EX: Kadokawa epubs)(see desc), small changes, and a few more safeguards
When upgrading all parts to a full volume, the full volume must still be an upgrade to the first part that is encountered in the existing library. So adding a parts regex to the ranked_keywords in settings.py that ranks a part volume lower than a regular volume would be needed. See my example below.

Keyword(r"(\b(Part)([-_. ]|)([0-9]+)\b)", -1)