Discord cover output requirements no longer needed (DESC)
The previous commits requirements can be discarded. The cover will be pulled automatically from the file, stored in memory, compressed in memory, and then sent to discord when sending a "New Volume Release" message. Absolutely nothing is required of the user now, it's all done automatically regardless of whether you have external covers or not.