Purged unncessary global keywords, adjusted bracket removal (desc) & chapter renaming logic, code cleanup, and wording changes
adjusted bracket removal: I improved the bracket removing logic to hopefully avoid removing a bracket from a series name that was apart of the series name. (currently this only runs on chapters, it was disabled on volumes due to some unexpected results, but volumes consistently have a volume keyword, so it's not really needed for them)

For example, (Cute) inside of the series name "The Strongest Swordsman Has Zero Equipment Slots, but He Can Equip up to 9999 Items if It's (Cute) Cursed Equipment":
"(Cute)" is a part of the name, and previously would've been removed. Now the code checks for an A-Z character on both sides, and will avoid removing it. This surely won't be perfect, and I will continue to adjust things in the future.