🐳 feat: Add Dockerfile for project deployment
This commit adds a Dockerfile to the project for easier deployment. The Dockerfile specifies the use of Python 3.11.0 and installs the necessary packages and requirements as defined in the `requirements.txt` file. It also creates a new user called "appuser" for security purposes and sets the ownership to appuser for the `/app` folder.