✨️ Change Docker login message for GitHub Actions
The Docker login step in the `docker-hub.yml` workflow file was updated to add the "/test" suffix to the "Log in to Docker Hub" action name. This change was made to make the name of the action more descriptive and informative. The change does not affect the functionality of the action.