✨ Refactor Dockerfile to use slim-buster image
The Dockerfile was modified to use the slim-buster image instead of the Alpine image for a smaller file size. The "addgroup" and "adduser" commands were replaced with "useradd" to simplify the creation of the "appuser". The image now sets ownership to appuser, switches to "appuser" and updates package lists before installing necessary packages and requirements. Finally, unnecessary files were removed.