added workflows  (#11)
* ✨ Add demo GitHub Action and .gitignore

The added GitHub Action is a demo that showcases how to run jobs on Ubuntu servers hosted by GitHub triggered by push events. The .gitignore file includes standard entries to exclude certain files from Git version control.

* ✨ Remove unnecessary command and update formatting

Removed the `no newline at end of file` and update formatting by removing unnecessary whitespace and replacing a period with a colon.

* 🐳 feat: Add Dockerfile for project deployment

This commit adds a Dockerfile to the project for easier deployment. The Dockerfile specifies the use of Python 3.11.0 and installs the necessary packages and requirements as defined in the `requirements.txt` file. It also creates a new user called "appuser" for security purposes and sets the ownership to appuser for the `/app` folder.

* ✨  Add GitHub Actions to Build and Push Docker Image

This change adds a GitHub Actions workflow 'docker-hub.yml' that builds and pushes a Docker image to Docker Hub whenever changes are pushed to the 'dev' or 'main' branches. The workflow includes three jobs: build, test, and push_to_registry. 'build' runs 'docker-compose' to build the Docker image from the Dockerfile. 'test' starts the Docker container and runs the tests. 'push_to_registry' logs in to Docker Hub, extracts metadata (tags, labels) for Docker, then builds and pushes the Docker image to Docker Hub. The workflow replaces the deleted 'github-actions-demo.yml' workflow.

* ✨️ Change Docker login message for GitHub Actions

The Docker login step in the `docker-hub.yml` workflow file was updated to add the "/test" suffix to the "Log in to Docker Hub" action name. This change was made to make the name of the action more descriptive and informative. The change does not affect the functionality of the action.

* Update README.md

* Main (#9)

* ✨ Add demo GitHub Action and .gitignore

The added GitHub Action is a demo that showcases how to run jobs on Ubuntu servers hosted by GitHub triggered by push events. The .gitignore file includes standard entries to exclude certain files from Git version control.

* ✨ Remove unnecessary command and update formatting

Removed the `no newline at end of file` and update formatting by removing unnecessary whitespace and replacing a period with a colon.

* 🐳 feat: Add Dockerfile for project deployment

This commit adds a Dockerfile to the project for easier deployment. The Dockerfile specifies the use of Python 3.11.0 and installs the necessary packages and requirements as defined in the `requirements.txt` file. It also creates a new user called "appuser" for security purposes and sets the ownership to appuser for the `/app` folder.

* ✨  Add GitHub Actions to Build and Push Docker Image

This change adds a GitHub Actions workflow 'docker-hub.yml' that builds and pushes a Docker image to Docker Hub whenever changes are pushed to the 'dev' or 'main' branches. The workflow includes three jobs: build, test, and push_to_registry. 'build' runs 'docker-compose' to build the Docker image from the Dockerfile. 'test' starts the Docker container and runs the tests. 'push_to_registry' logs in to Docker Hub, extracts metadata (tags, labels) for Docker, then builds and pushes the Docker image to Docker Hub. The workflow replaces the deleted 'github-actions-demo.yml' workflow.

* ✨️ Change Docker login message for GitHub Actions

The Docker login step in the `docker-hub.yml` workflow file was updated to add the "/test" suffix to the "Log in to Docker Hub" action name. This change was made to make the name of the action more descriptive and informative. The change does not affect the functionality of the action.

* Add workflow to build and push Docker image to Docker Hub

* Add workflow to build and push Docker image to Docker Hub to use dockerFile instead

* Add workflow to build and push Docker image to Docker Hub to use dockerFile instead

* Add workflow to build and push Docker image to Docker Hub to use dockerFile instead

* Add workflow to build and push Docker image to Docker Hub to use dockerFile instead

* Add workflow to build and push Docker image to Docker Hub to use dockerFile instead

* Add workflow to build and push Docker image to Docker Hub to use dockerFile instead

* Add workflow to build and push Docker image to Docker Hub to use dockerFile instead

* Add workflow to build and push Docker image to Docker Hub to use dockerFile instead

* Add workflow to build and push Docker image to Docker Hub to use dockerFile instead

* Merge main branch into dev branch

* 🐳 Use Alpine image in Dockerfile
Switched from Debian Buster slim to Alpine 3.15 to reduce image size and improve security.

- Replaced Debian Buster slim with Alpine 3.15 to reduce image size and improve security.
- Installed unrar and tzdata through apk package manager instead of Debian's apt-get command.
- Removed unused dependencies using the apk command instead of apt-get autoremove and rm command.
- Replaced groupadd, useradd, chown, and chmod commands with addgroup, adduser, and chown commands to simplify user management.
- Updated the Dockerfile to use the above changes.

* ✨ Refactor Dockerfile to use slim-buster image

The Dockerfile was modified to use the slim-buster image instead of the Alpine image for a smaller file size. The "addgroup" and "adduser" commands were replaced with "useradd" to simplify the creation of the "appuser". The image now sets ownership to appuser, switches to "appuser" and updates package lists before installing necessary packages and requirements. Finally, unnecessary files were removed.

* Changes to be committed:
	new file:   test

Changes not staged for commit:
	deleted:    test

* deleted:    test

* Revert "Main (#9)"

This reverts commit eb6c04a9fba8ea8716c5ff4802cbdc86e85e1209.

* Update docker-hub.yml

---------

Co-authored-by: Zach Stultz <hcazs4@live.com>