🐳 Optimize Dockerfile multi-stage build & 🚚 Remove unnecessary dependencies
Optimized the Dockerfile to better leverage multi-stage builds, resulting in a cleaner and more efficient image. Unneeded dependencies such as unrar-free, tzdata, nano, and rclone were moved from the build stage to the final stage to simplify the build process and reduce the size of the intermediate images. Also, the build process now involves compiling the requirements in one step within the app's context to ensure correct dependency resolution. The entrypoint was modified to launch the necessary scripts, adhering to best practices for container initialization.